requirements.txt
# Video-to-SOP Generator Requirements

# Video Processing
opencv-python>=4.8.0
moviepy>=1.0.3
numpy>=1.24.0

# Audio Processing (Whisper via Groq)
groq>=0.4.0

# AI/LLM
google-generativeai>=0.8.5
python-dotenv>=1.0.0

# PDF Generation
reportlab>=4.0.0
Pillow>=10.0.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
